+++ src/index.ts
@@ -1,7 +1,11 @@
 import RulesResolver from './RulesResolver';
 import { Rule, Stylesheet } from 'css';
 import { MediaQueryInterface } from './MediaQueryInterface';
-import { lint } from './lint';
+import { lint, LintRuleType } from './lint';
+
+const process = (ctx: Stylesheet): Array<Rule | MediaQueryInterface> => {
+  return new RulesResolver(ctx.stylesheet.rules).resolve();
+};
 
 export const preprocess: Function = (ctx: Stylesheet, next: Function): Function => {
   ctx.stylesheet.rules = process(ctx);
@@ -9,11 +13,9 @@ export const preprocess: Function = (ctx: Stylesheet, next: Function): Function 
   return next();
 };
 
-export const process = (ctx: Stylesheet): Array<Rule | MediaQueryInterface> => {
-  return new RulesResolver(ctx.stylesheet.rules).resolve();
-};
-
-export const processLint = (ctx: Stylesheet): Array<string> => {
+export const processLint = (
+  ctx: Stylesheet,
+): { rules: Array<LintRuleType>; isValid: boolean; isBemDetected: boolean } => {
   ctx.stylesheet.rules = process(ctx);
 
   return lint(ctx);